requirements.txt
selenium==4.15.2
pandas==2.1.4
python-dotenv==1.0.0